+++ backend/src/management/order.rs
@@ -18,7 +18,7 @@ pub fn create_order(
     offramper_providers: HashMap<PaymentProviderType, PaymentProvider>,
     blockchain: Blockchain,
     token: Option<String>,
-    crypto_amount: u64,
+    crypto_amount: u128,
     offramper_address: TransactionAddress,
 ) -> Result<u64> {
     let order = Order::new(
